(build) prototype inconsistencies compiler warnings
FossilOrigin-Name: f3eb4ae3ac458f2f6b80f50150969b8eff9fb691493106804273c44f9614d098

+++ src/frameserver/decode/default/decode_av.c
@@ -256,7 +256,7 @@ static void audio_play(void *data,
  * playback engine resynch as part of the _RESET */
 	 if (decctx.shmcont.abufsize <= decctx.shmcont.abufused){
 			LOG("shmif_cont-rejecting:size=%zu:used:%zu",
-				decctx.shmcont.abufsize, decctx.shmcont.abufused);
+				(size_t)decctx.shmcont.abufsize, (size_t)decctx.shmcont.abufused);
 			arcan_shmif_unlock(&decctx.shmcont);
 		}
